+++ tests/test_settings.py
@@ -2668,7 +2668,7 @@ def main() -> None:
     defaults = {}
     setting_ids = set()
 
-    assert data.version == "2.9.0"
+    assert data.version == "2.9.1"
 
     gradient_name = localization["mod_name"]["en"]
     assert gradient_name.startswith("{#color(174,239,105)}B")